Rename package configUtil to config.
Reason:The package names should be single word!
if there is a necessity to combine two words 'config' and 'util', the convention is 'config.util'
Here 'util' is not necessary so dropping it
