+++ sparkler-api/src/main/java/edu/usc/irds/sparkler/config/PluginsProps.java
@@ -1,4 +1,4 @@
-package edu.usc.irds.sparkler.configUtils;
+package edu.usc.irds.sparkler.config;
 
 
 import java.util.Map;
