+++ sparkler-plugins/urlfilter-regex/src/main/java/edu/usc/irds/sparkler/config/UrlFilterProps.java
@@ -1,4 +1,4 @@
-package edu.usc.irds.sparkler.configUtils;
+package edu.usc.irds.sparkler.config;
 
 
 import edu.usc.irds.sparkler.BaseConfig;
