+++ sparkler-plugins/urlfilter-regex/src/test/java/edu/usc/irds/sparkler/config/UrlFilterPropsTest.java
@@ -1,4 +1,4 @@
-package edu.usc.irds.sparkler.configUtils;
+package edu.usc.irds.sparkler.config;
 
 import edu.usc.irds.sparkler.Constants;
 import edu.usc.irds.sparkler.SparklerConfig;
